Removed warnings
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/aims/aims-free/trunk@62960 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ aimsdata/src/tests/AimsStructuringElement_test/structuringelement_test.cc
@@ -73,10 +73,10 @@ int main( int argc, const char** argv )
     app.initialize();
     
     StructuringElement se = StructuringElement::getStructuringElement( type, 
-                                                                               Point3d(amplitude, 
-                                                                                       amplitude, 
-                                                                                       amplitude),
-                                                                               amplitude );
+                                                                       Point3d((int16_t)amplitude, 
+                                                                               (int16_t)amplitude, 
+                                                                               (int16_t)amplitude),
+                                                                       amplitude );
 
     AimsData<int16_t> out( (int)(2 * amplitude + 1),
                            (int)(2 * amplitude + 1),
